+++ src/Appeaser/IAsyncQuery.cs
@@ -1,0 +1,4 @@
+namespace Appeaser
+{
+    public interface IAsyncQuery<TReturn> { }
+}