+++ src/Appeaser/IMediator.cs
@@ -1,8 +1,13 @@
-﻿namespace Appeaser
+﻿using System.Threading.Tasks;
+
+namespace Appeaser
 {
     public interface IMediator
     {
         TResponse Request<TResponse>(IQuery<TResponse> query);
+        Task<TResponse> Request<TResponse>(IAsyncQuery<TResponse> query);
+
         TResult Send<TResult>(ICommand<TResult> query);
+        Task<TResult> Send<TResult>(IAsyncCommand<TResult> query);
     }
 }