+++ src/Appeaser/Mediator.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Diagnostics;
 using System.Reflection;
+using System.Threading.Tasks;
 using Appeaser.Exceptions;
 
 namespace Appeaser
@@ -8,23 +9,20 @@ namespace Appeaser
     public class Mediator : IMediator
     {
         protected readonly IMediatorHandlerFactory HandlerFactory;
-        protected readonly Type OpenQueryHandlerType = typeof(IQueryHandler<,>);
-        protected readonly Type OpenCommandHandlerType = typeof(ICommandHandler<,>);
+        protected readonly IMediatorSettings Settings;
 
-        public Mediator(IMediatorHandlerFactory handlerFactory)
+        public Mediator(IMediatorHandlerFactory handlerFactory, IMediatorSettings settings = null)
         {
             HandlerFactory = handlerFactory;
+            Settings = settings ?? new MediatorSettings();
         }
 
-        [DebuggerStepThrough]
+        //[DebuggerStepThrough]
         public virtual TResponse Request<TResponse>(IQuery<TResponse> query)
         {
             try
             {
-                var requestType = typeof(TResponse);
-                var queryType = query.GetType();
-                var requestingHandlerType = OpenQueryHandlerType.MakeGenericType(queryType, requestType);
-                var handler = HandlerFactory.GetHandler(requestingHandlerType);
+                var handler = GetHandler<TResponse>(typeof(IQueryHandler<,>), query);
                 if (handler == null)
                 {
                     throw new MediatorQueryException("No query handler of type {0} could be found", query.GetType());
@@ -34,7 +32,31 @@ namespace Appeaser
             }
             catch (Exception ex)
             {
-                if (ex is MediatorQueryException)
+                if (ex is MediatorQueryException || !Settings.WrapExceptions)
+                {
+                    throw;
+                }
+
+                throw new MediatorQueryException(ex, query.GetType());
+            }
+        }
+
+        [DebuggerStepThrough]
+        public virtual async Task<TResponse> Request<TResponse>(IAsyncQuery<TResponse> query)
+        {
+            try
+            {
+                var handler = GetHandler<TResponse>(typeof(IAsyncQueryHandler<,>), query);
+                if (handler == null)
+                {
+                    throw new MediatorQueryException("No query handler of type {0} could be found", query.GetType());
+                }
+
+                return await InvokeHandlerAsync<TResponse>(handler, query);
+            }
+            catch (Exception ex)
+            {
+                if (ex is MediatorQueryException || !Settings.WrapExceptions)
                 {
                     throw;
                 }
@@ -48,10 +70,7 @@ namespace Appeaser
         {
             try
             {
-                var returnType = typeof(TResult);
-                var commandType = command.GetType();
-                var requestingHandlerType = OpenCommandHandlerType.MakeGenericType(commandType, returnType);
-                var handler = HandlerFactory.GetHandler(requestingHandlerType);
+                var handler = GetHandler<TResult>(typeof(ICommandHandler<,>), command);
                 if (handler == null)
                 {
                     throw new MediatorCommandException("No command handler of type {0} could be found", command.GetType());
@@ -61,7 +80,7 @@ namespace Appeaser
             }
             catch (Exception ex)
             {
-                if (ex is MediatorCommandException)
+                if (ex is MediatorCommandException || !Settings.WrapExceptions)
                 {
                     throw;
                 }
@@ -71,10 +90,64 @@ namespace Appeaser
         }
 
         [DebuggerStepThrough]
+        public virtual async Task<TResult> Send<TResult>(IAsyncCommand<TResult> command)
+        {
+            try
+            {
+                var handler = GetHandler<TResult>(typeof(IAsyncCommandHandler<,>), command);
+                if (handler == null)
+                {
+                    throw new MediatorCommandException("No command handler of type {0} could be found", command.GetType());
+                }
+
+                return await InvokeHandlerAsync<TResult>(handler, command);
+            }
+            catch (Exception ex)
+            {
+                if (ex is MediatorCommandException || !Settings.WrapExceptions)
+                {
+                    throw;
+                }
+
+                throw new MediatorCommandException(ex, command.GetType());
+            }
+        }
+
+        [DebuggerStepThrough]
+        protected virtual object GetHandler<TResponse>(Type handlerType, object parameter)
+        {
+            var requestType = typeof(TResponse);
+            var parameterType = parameter.GetType();
+            var requestingHandlerType = handlerType.MakeGenericType(parameterType, requestType);
+            return HandlerFactory.GetHandler(requestingHandlerType);
+        }
+
+        [DebuggerStepThrough]
         protected virtual TReturn InvokeHandler<TReturn>(object handler, object parameter)
         {
             var method = handler.GetType().GetRuntimeMethod("Handle", new[] { parameter.GetType() });
-            return (TReturn)method.Invoke(handler, new[] { parameter });
+            try
+            { 
+                return (TReturn)method.Invoke(handler, new[] { parameter });
+            }
+            catch(TargetInvocationException ex)
+            {
+                throw ex.InnerException;
+            }
+        }
+
+        [DebuggerStepThrough]
+        protected virtual async Task<TReturn> InvokeHandlerAsync<TReturn>(object handler, object parameter)
+        {
+            var method = handler.GetType().GetRuntimeMethod("Handle", new[] { parameter.GetType() });
+            try
+            { 
+                return await(Task<TReturn>)method.Invoke(handler, new[] { parameter });
+            }
+            catch (TargetInvocationException ex)
+            {
+                throw ex.InnerException;
+            }
         }
     }
 }