+++ src/Appeaser/MediatorSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Appeaser
+{
+    public class MediatorSettings : IMediatorSettings
+    {
+        public bool WrapExceptions => true;
+    }
+}